+++ packages/f-header/src/components/Navigation.vue
@@ -50,7 +50,7 @@
                 </li>
 
                 <li
-                    :class="['c-nav-list-item has-sublist', { 'is-hidden': !userInfo.isAuthenticated, 'open': navIsOpen }]"
+                    :class="['c-nav-list-item has-sublist', { 'is-hidden': !userInfo, 'open': navIsOpen }]"
                     @mouseover="openNav"
                     @mouseleave="closeNav"
                     @keyup.esc="closeNav">
@@ -117,7 +117,7 @@
                 </li>
 
                 <li
-                    v-if="!userInfo.isAuthenticated"
+                    v-if="!userInfo"
                     class="c-nav-list-item"
                     data-js-test="login">
                     <a
@@ -152,7 +152,7 @@
                     </li>
 
                     <li
-                        v-if="userInfo.isAuthenticated"
+                        v-if="userInfo"
                         class="c-nav-list-item"
                         data-js-test="logout">
                         <a
@@ -242,13 +242,18 @@ export default {
         userInfoUrl: {
             type: String,
             default: '/api/account/details'
+        },
+        orderCountUrl: {
+            type: String,
+            default: '/api/analytics/ordercount'
         }
     },
     data () {
         return {
             navIsOpen: false,
             currentScreenWidth: 0,
-            userInfo: this.userInfoProp
+            userInfo: this.userInfoProp,
+            orderCountInfo: false
         };
     },
     computed: {
@@ -268,7 +273,7 @@ export default {
     },
     mounted () {
         if (!this.userInfo) {
-            this.setUserInfo();
+            this.fetchUserInfo();
         }
         this.currentScreenWidth = window.innerWidth;
         window.addEventListener('resize', throttle(this.onResize, 100));
@@ -289,29 +294,93 @@ export default {
         onResize () {
             this.currentScreenWidth = window.innerWidth;
         },
+        // When hamburger menu is clicked we want to trigger toggling of navigation and emit the state to the parent to add transparent class
+        onHamburgerMenuClick () {
+            this.onNavToggle();
+            this.$emit('onMobileNavToggle', this.navIsOpen);
+        },
         // If userInfoProp wasn't passed we make a call for userInfo on mounted hook
-        async setUserInfo () {
+        async fetchUserInfo () {
             try {
                 const { data } = await axios.get(this.userInfoUrl, {
                     headers: {
                         credentials: 'same-origin'
                     }
                 });
-                if (data) {
+                if (data.isAuthenticated) {
                     this.userInfo = data;
+                    this.saveUserData();
+                } else {
+                    this.userInfo = false;
                 }
             } catch (err) {
                 if (this.justLog.error) {
-                    this.justLog.error('Error handling "setUserInfo" action', err);
+                    this.justLog.error('Unable to get user information from "fetchUserInfo"', err);
                 }
             }
         },
-        // When hamburger menu is clicked we want to trigger toggling of navigation
-        // + emit the state of `navIsOpen` attr to the header component to change header styles in case of transparency
-        // in open nav state mobile header should become white, not transparent
-        onHamburgerMenuClick () {
-            this.onNavToggle();
-            this.$emit('onMobileNavToggle', this.navIsOpen);
+        // fetches the order count for the user
+        async fetchOrderCountAndSave () {
+            try {
+                const { data } = await axios.get(this.orderCountUrl, {
+                    headers: {
+                        credentials: 'same-origin'
+                    }
+                });
+                if (data) {
+                    this.orderCountInfo = data;
+                    this.setAnanlyticsBlob();
+                    this.enrichUserDataWithCount(data);
+                    this.pushUserData();
+                }
+            } catch (err) {
+                if (this.justLog.error) {
+                    this.justLog.error('Unable to get order count from "fetchOrderCountAndSave', err);
+                }
+            }
+        },
+        // Sets the order count info in local storage
+        setAnanlyticsBlob () {
+            window.localStorage.setItem('je-analytics', JSON.stringify(this.orderCountInfo));
+        },
+        // Gets the order count info in local storage
+        getLocalAnalyticsBlob () {
+            return window.localStorage.getItem('je-analytics');
+        },
+        // Updates the user information with the count
+        enrichUserDataWithCount () {
+            this.userInfo.orderCount = this.orderCountInfo.Count;
+        },
+        // Pushes the user info to the windows data layer
+        pushUserData () {
+            window.dataLayer.push(this.userInfo);
+        },
+        // Checks that the element exist's
+        orderCountSupported () {
+            const supportedEl = document.querySelector('[data-order-count-supported]');
+            if (supportedEl && supportedEl.value) {
+                // Case insensitive regex test for value="true"
+                return /^true$/i.test(supportedEl.value);
+            }
+            return false;
+        },
+        // Saves the user data and calls the nessesary methods based on what is already there
+        saveUserData () {
+            const localOrderCount = JSON.parse(this.getLocalAnalyticsBlob);
+            const currentTime = new Date().getTime();
+            const localOrderCountExpires = Date.parse(localOrderCount.Expires);
+
+            if (!this.orderCountSupported()) {
+                this.pushUserData();
+            }
+            if (!this.getLocalAnalyticsBlob) {
+                this.fetchOrderCountAndSave();
+            }
+            if (localOrderCountExpires < currentTime) {
+                this.fetchOrderCountAndSave();
+            }
+            this.enrichUserDataWithCount(localOrderCount);
+            this.pushUserData();
         }
     }
 };