+++ packages/f-services/CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 0.9.0
 ------------------------------
-*September 30, 2019*
+*October 1, 2019*
 
  ### Added
 - `lodash-es` throttle for the `addEvent` service